Stabilize embedded PostgreSQL test setup
Bump `pg-embed-setup-unpriv` to `0.5.1` and refresh `Cargo.lock`
so the worker subprocess receives the parent environment in CI. Pin the
embedded PostgreSQL runtime version in the Rust test environment to avoid
release-listing requests during archive resolution.

Extract audit reporting helpers from `security/audit-utils.js` so the
shared audit utility module stays below the repository file-size limit.
