+++ backend/Cargo.toml
@@ -22,7 +22,7 @@ clap = { version = "4", features = ["derive"] }
 
 async-trait = "0.1.89"
 color-eyre = "0.6.5"
-pg_embedded_setup_unpriv = { package = "pg-embed-setup-unpriv", version = "0.5.0" }
+pg_embedded_setup_unpriv = { package = "pg-embed-setup-unpriv", version = "0.5.1" }
 postgresql_embedded = { version = "0.20.0", features = ["tokio"] }
 postgres = { version = "0.19.12", features = ["with-uuid-1"] }
 paste = "1.0.15"
